Candidate Search and Saved are functional
both webpages work functionally and have messages that will display if there is no more profiles in the que or if there are no saved profiles. However with this in mind, the API is coded to render a new array of profiles once the first array is empty, so seeing the there are no more candidates display message is, in theory, not possible. However, this API is slow and glitchy so it may display in between renders.